server/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
click==8.1.6
dnspython==2.4.1
exceptiongroup==1.1.2
fastapi==0.100.1
h11==0.14.0
idna==3.4
numpy==1.25.1
pydantic==2.1.1
pydantic_core==2.4.0
pymongo==4.4.1
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
urllib3==1.26.6
uvicorn==0.23.1
